Add spell check process to guides
To prevent easily made spelling errors which should also help improve the quality of the guides, this change adds a `.spelling` file which can be used in conjunction with [mdspell](https://www.npmjs.com/package/markdown-spellcheck) to check for spelling errors in the markdown files.

`.spelling` is used to maintain a list of additional words that should be checked, or words that can be ignored. This change also includes initial spelling corrections as a result of running mdspell, and an update to CONTRIBUTION.md with instructions on how to use mdspell.