Add Travis config yml
Though not a development project in the normal sense, we would still like to use CI to ensure quality on the contributions and changes made to it.

This change adds a `.travis.yml` to the project currently with the intention of checking contributions for spelling errors. Also included is an update to the instructions on how best to run [markdown-spellcheck](https://www.npmjs.com/package/markdown-spellcheck) locally.