Remove QA step from process
Having read @reggieb comment regards the process as outlined at the time, it became clear that the goal of focused PR's related to single changes was conflicting with a QA step that is there to check the overarching story is complete.

Our QA is focused on the overall feature, most likely defined in a Jira story as mentioned in the comment. They are not interested in the sequence of changes which delivered that feature.

Therefore this change removes the QA step from the process. Now after the code review is complete and the commits have been squashed, the contributor goes straight to merging in the PR.