Simplify text and use better examples
To try and make the process clearer this changes aims to use clearer, less verbose explanations of what to do. It has also removed text that simply expanded existing points or that were not directly related to the process.

Finally the example commit message has been updated with one that better represents a specific, focused commit as outlined in the goals.