Remove markdown-spellcheck from build
It has recently been found by the team that **markdown-spellcheck** does not ignore text in backticks. This tends to be code so is not expected to to be recognised text, but **markdown-spellcheck** still flags it and forces you to go through each false-positive using its interactive command line process.

Till a better alternative can be found, or until we can add this option to ignore text in backticks to the package we need to remove it from the build process. The guides and documents we write tend to be 'tech' heavy and feature loads of code examples, and having to deal with this issue is more likely to dissuade team members from writing documentation, something we strongly wish to avoid.

The changes include in this are

- Remove call to markdown-spellcheck from travis.yml
- Switch travis.yml to be a ruby build rather than node build
- Update contribution guidance to recommend people run the tool, but not require it