Add missing links and README's to repo
In PR's #9 and #13 new guides were added but the root README was not updated with links to them, and README's were not placed in the roots of some of the new folders.

Having links to all the guides on the root README aids visibility of the content, and having a README in each folder ensures consistency and an opportunity to provide details to the purpose of a section of the guides.

This changes adds the missing links and README's.