Update PR guide to use labels rather emojis
Initial feedback on the new PR process from the team was that it was difficult to see which PR's needed reviewing. It was suggested that using the labels provided by GitHub rather than comments with emoji's would make it clearer and easier to see which PR's needed reviewing.

This change updates the PR process to use labels, but also includes using labels to cover

- ready to review
- in review
- ship it