Simplify the folder structure for guides
The current folder structure is based on what [ThoughtBot](https://github.com/thoughtbot/guides) did for theirs. It was assumed that essentially each document needed to be in its own folder and named README.md but since then have seen this is not necessary. We can have multiple documents in a folder each named after what they are which should make things simpler.

This change updates the folder structure and names of the files to simplify the dst-guides project from a user and maintainer's point of view.