Simplification of the 'services' folder
This change removes using a separate folder for each sub topic in a service to use a named file at the service root. For example in `/services/WEX` we have removed the `state-engine` folder, moved the state-engine guide it contained to the root of WEX, renameing it `state_engine.md`.