Make guides applicable to all Digital Services
When guides was first put together it was intended for the use of the Digital Service Team (south) only, though made public so all could access it for reference.

At least within the Regulated Customer Digital Programme the intention is for what is listed in here to become the defacto standard, with an eye for it applying to all digital services going forward.

Therefore this change is to update the root README and any other relevant guides to be *generic* to all DS's rather than *specific* to DST.