Add initial guide covering CI for ruby services
This first version is intended to give a high level view of the setup of CI for ruby based projects. It covers how GitHub should be setup and which services you are expected to integrate to, and goes onto to cover what we do in our Jenkins environments.

Currently readers would need to refer to existing examples to actually implement what is discussed in this guide. The intention with this version is to help them understand what they need to be doing.