Add link to supportive blog on coding in the open
Recently (at the time of writing) the Ministry of Justice published a blog post on [Why they code in the open](https://mojdigital.blog.gov.uk/2017/02/21/why-we-code-in-the-open/). This is not just evidence of another department making code open, but believing that writing code in the open makes for **better code**. This is the main argument behind our principle of **Work in the open**.

The blog post also has some great responses for one of the most common arguments against this; Doesn’t it give attackers an advantage? Therefore it was felt advantageous to add a link to it in the [Work in the open](/principles/work_in_the_open.md) guide.