Add new principle 'Code is not yours'
This adds a second proposed principle; that when writing code, commit messages and documentation we should be thinking of those that come after us and not ourselves. Hence the title **The code is not 'yours'**.

The hope is this leads to more consistency in our projects, makes them easier to maintain, and reduces the time it takes to make future changes.

It also includes some tweaks to the Pull request process to support this goal, and a couple of fixes.