Add new principle 'Master is always shippable'
This adds a third proposed principle; that the main branch on any project (typically master) is always in a state of ready to ship to production.

The aim is to provide consistency in what to expect when looking at a project; essentially the separation between what a team considers done and what they are still working on. As a potential user, I should always be able to clone or fork a project and know that all tests will pass and that it should run (assuming all dependencies are met) when the main branch is checked out.

One of the inspirations for this was a confrontation with a fellow developer. For 'reasons' I'd taken an interest in another project and began to look at the code and information provided. Seeing possible issues on `master` I made some comments, only for them to be rejected on the basis this was a project in its early stages and still being worked. The team was not seeking outside comments on work that was still incomplete.

My hope is this principle will prevent this type of misunderstanding in the future.