Add guide on all things related to GitHub use
Our principles push for us [Work in the
open](https://github.com/DEFRA/dst-guides/blob/master/principles/work_in_the_open.md),
which means publishing stuff to **GitHub**. However we don't actually
have documented anywhere what the requirements are to access it, how we
have structured it, and what we expect of users once they have access to
our organisation on **GitHub**.

Part of the reason was that we had not formally agreed some of these
decisions, but this has been rectified following a meeting of the core
maintainers.

So this guide captures the decisions made, and forms a basis to
colloborate and discuss how we want to use GitHub in the future.

Changes include
- Create a new 'GitHub' section
- New guide on GitHub usage
- Moved 'New projects' guide to GitHub section
- Update root README to reflect changes