Add details on the PAFS service to guides
PAFS is a service in public beta and CI work is due to start on it soon. This change adds an initial guide to provide basic details on the service. This should act as a base for those working on the service to add more information about it.

Till then, we acknowledge it's existance in DST-guides with this new guide.