Add details on the Waste carriers service 2 guides
Waste carriers registration is a service in public beta. This change adds an initial guide to provide basic details on the service. This should act as a base for those working on the service to add more information about it future.

Till then, we acknowledge it's existence in DST-guides with this new guide.