Update .rubocop.yml and state 2.4 as our ruby ver.
The version of the .rubocop.yml here was from the waste exemptions project and did not reflect how we on the 'small team' no configure rubocop. It also was incomptable with the latest versions of rubocop which have changed the namespaces of some of the cops.

So we have updated it to reflect what we are currently using, but would welcome feedback on what is set out.

Finally where we previously recommended simply using the very latest versions of ruby and rails, with no new projects intending to be done in RoR there is limited support for the existing projects.

Therefore we're trying to standardise on keeping everything up to date on the latest 2.4 version of ruby, and version 4 of Rails. This is till such time as the projects are re-written, or we have the scope to migrate all to Rails 5.