Update Ruby-based CI guidance (#53)
Deppbot has not been pushing updates for several months now. We've been using Dependabot on the waste carriers renewals project and have found this is a good substitute (with some configuration).

DependencyCI has merged into Tidelift and now follows a subscription model, so the features we used to use are now only available with a paid subscription. So for that reason I've removed it from the guidance alltogether.