Update pull request process to make it simpler (#55)
When we first started doing pull requests, the GitHub UI was not as feature rich as it is now, and it was a new process to some of us using it.

Since then GitHub has added new functionality, and those of us using PR's have increased our experience.

The result is a more simplified PR process. For example gone is the need to create custom labels and assign them to the PR during its various stages. Now simply request reviews and await an approval. Also its no longer necessary to handle squashing your commits locally. This can be done from the UI when you complete and merge the PR.

So these changes update the PR process to reflect what those of us using PR's actually do, along with simplifying it in the hopes that more make use of them.