Add a teams folder (#65)
Recently in the Ruby services development team we wanted to start documenting some of the decisions we have made about the ways we work.

We use Google docs as a team, but the kind of stuff we like to document tends to feature code examples and Google docs isn't great for that. Turns out it's also not great for working with Markdown, which is great for code examples!

So we decided that there shouldn't be anything wrong with creating a space on DST-Guides for teams to start recording stuff unique to them. It fits with trying to make DST-Guides a one-stop-shop for new developers, and if other teams contribute would be a way of collaborating and sharing ways of working.